Add explicit table creation param

+++ gen_field_data.pl
@@ -34,12 +34,14 @@ my $man = 0;
 my %ignore_fields;
 my $nocache = 0;
 my $insert_table = undef;
+my $create_table = 0;
 
 GetOptions(
     'db=s%' => sub { my $onam = $_[1]; my $oval = $_[2]; if (defined($dbdata{$onam})) { $dbdata{$onam} = $oval; } else { die("Unknown db setting."); } },
     'ignore=s' => sub { my ($onam, $oval) = @_; foreach my $tmp (split/,/, $oval) { $ignore_fields{$tmp} = 1; } },
     'nocache' => \$nocache,
     'insert=s' => \$insert_table,
+    'create' => \$create_table,
     'help|h|?' => \$help,
     'man' => \$man
     ) or pod2usage(2);
@@ -75,8 +77,10 @@ $dbh->do("SET NAMES 'utf8';");
 
 if (defined($insert_table)) {
     my @tmparr = ($insert_table);
-    $dbh->do("DROP TABLE IF EXISTS $insert_table");
-    $dbh->do("CREATE TABLE $insert_table (bibnum int unsigned not null, idx int unsigned not null, tag varchar(3) not null, subfield varchar(1), ind1 varchar(1), ind2 varchar(2), value varchar(1024), KEY valueidx (value), KEY tagidx (tag), KEY fieldidx (tag, subfield))");
+    if ($create_table) {
+	$dbh->do("DROP TABLE IF EXISTS $insert_table");
+	$dbh->do("CREATE TABLE $insert_table (bibnum int unsigned not null, idx int unsigned not null, tag varchar(3) not null, subfield varchar(1), ind1 varchar(1), ind2 varchar(2), value varchar(1024), KEY valueidx (value), KEY tagidx (tag), KEY fieldidx (tag, subfield))");
+    }
     $ins_sth = $dbh->prepare("INSERT INTO $insert_table (bibnum, idx, tag, subfield, ind1, ind2, value) values (?, ?, ?, ?, ?, ?, ?)");
 }
 
@@ -169,6 +173,11 @@ and dumping it out at end of script.
 Insert data into table `tablename` instead of printing it to
 stdout. Drops and recreates the table automatically. Uses
 the same database settings as reading the data.
+
+=item B<-create>
+
+Drop and create the table defined with -insert before inserting
+data into it.
 
 =item B<-ignore=fieldspecs>
